Renamed LogIntegration to S3LogIntegration + screen LogSourceOnboarding
