--- web/src/components/wizards/LogSourceWizard/StackDeploymentPanel/graphql/getLogCfnTemplate.graphql
@@ -1,6 +1,0 @@
-query GetLogCfnTemplate($input: GetLogIntegrationTemplateInput!) {
-  getLogIntegrationTemplate(input: $input) {
-    body
-    stackName
-  }
-}
