+++ web/src/components/wizards/S3LogSourceWizard/StackDeploymentPanel/graphql/getLogCfnTemplate.graphql
@@ -1,0 +1,6 @@
+query GetLogCfnTemplate($input: GetS3LogIntegrationTemplateInput!) {
+  getS3LogIntegrationTemplate(input: $input) {
+    body
+    stackName
+  }
+}
