--- web/src/graphql/fragments/LogIntegrationDetails.graphql
@@ -1,23 +1,0 @@
-fragment LogIntegrationDetails on LogIntegration {
-  integrationId
-  integrationLabel
-  awsAccountId
-  createdAtTime
-  createdBy
-  kmsKey
-  s3Bucket
-  s3Prefix
-  logTypes
-  stackName
-  health {
-    processingRoleStatus {
-      ...IntegrationItemHealthDetails
-    }
-    s3BucketStatus {
-      ...IntegrationItemHealthDetails
-    }
-    kmsKeyStatus {
-      ...IntegrationItemHealthDetails
-    }
-  }
-}
