+++ web/src/helpers/utils.tsx
@@ -31,6 +31,7 @@ import {
   INCLUDE_SPECIAL_CHAR_REGEX,
   INCLUDE_UPPERCASE_REGEX,
   CHECK_IF_HASH_REGEX,
+  SOURCE_LABEL_REGEX,
 } from 'Source/constants';
 import mapValues from 'lodash-es/mapValues';
 import sum from 'lodash-es/sum';
@@ -57,6 +58,12 @@ export const createYupPasswordValidationSchema = () =>
     .matches(INCLUDE_LOWERCASE_REGEX, 'Include at least 1 lowercase character')
     .matches(INCLUDE_UPPERCASE_REGEX, 'Include at least 1 uppercase character')
     .matches(INCLUDE_SPECIAL_CHAR_REGEX, 'Include at least 1 special character');
+
+export const integrationLabelValidation = () =>
+  Yup.string()
+    .required()
+    .matches(SOURCE_LABEL_REGEX, 'Can only include alphanumeric characters, dashes and spaces')
+    .max(32, 'Must be at most 32 characters');
 
 /**
  * checks whether the input is a valid UUID
