+++ web/src/pages/CreateLogSource/CreateS3LogSource/graphql/addS3LogSource.graphql
@@ -1,0 +1,5 @@
+mutation AddS3LogSource($input: AddS3LogIntegrationInput!) {
+  addS3LogIntegration(input: $input) {
+    ...S3LogIntegrationDetails
+  }
+}
