+++ web/src/pages/EditS3LogSource/graphql/getS3LogSource.graphql
@@ -1,0 +1,5 @@
+query GetS3LogSource($id: ID!) {
+  getS3LogIntegration(id: $id) {
+    ...S3LogIntegrationDetails
+  }
+}
